清理代码告警信息

+++ frameworks/native/audioutils/src/audio_utils.cpp
@@ -714,7 +714,7 @@ bool SignalDetectAgent::DetectSignalData(int32_t *buffer, size_t bufferLen)
             rightZeroSignal = index + 1;
             hasNoneZero = true;
             if (currentPeakIndex == -1 || tempMax > currentPeakSignal) {
-                currentPeakIndex = index;
+                currentPeakIndex = static_cast<ssize_t>(index);
                 currentPeakSignal = tempMax;
             }
         }
@@ -755,7 +755,7 @@ bool AudioLatencyMeasurement::MockPcmData(uint8_t *buffer, size_t bufferLen)
     memset_s(buffer, bufferLen, 0, bufferLen);
     int16_t *signal = signalData_.get();
     size_t newlyMocked = bufferLen * MILLISECOND_PER_SECOND /
-        (channelCount_ * sampleRate_ * formatByteSize_);
+        (uint32_t)(channelCount_ * sampleRate_ * formatByteSize_);
     mockedTime_ += newlyMocked;
     if (mockedTime_ >= MOCK_INTERVAL) {
         mockedTime_ = 0;
