+++ services/audio_service/client/src/renderer_in_client.cpp
@@ -1126,7 +1126,7 @@ int32_t RendererInClientInner::SetAudioEffectMode(AudioEffectMode effectMode)
 
 int64_t RendererInClientInner::GetFramesWritten()
 {
-    return totalBytesWritten_ / sizePerFrameInByte_;
+    return totalBytesWritten_ / static_cast<ssize_t>(sizePerFrameInByte_);
 }
 
 int64_t RendererInClientInner::GetFramesRead()
